argDefaultVal {argparse}	R Documentation

Return default value of an optional argument

Description

Given a parsed parameter table, returns the default value of the argument with the given name. Returns NULL if the named argument is mandatory.

Usage

argDefaultVal(argName, expectedArgs)

Arguments

argName
String with the argument name.

expectedArgs
parsed list of expected Arguments, as returned by parseExpectedArgs.

Value

Default value, type of the named argument.

Author(s)

Moritz Lell

See Also

parseExpectedArgs, argType, argNames

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (argName, expectedArgs) 
{
    stopifnot("expectedArgs" %in% class(expectedArgs))
    stopifnot(mode(argName) == "character")
    val = expectedArgs[[argName]]$val
    if (is.null(val)) 
        return(NULL)
    if (is.na(val)) 
        return(NA)
    if (val == attr(expectedArgs, "expectedMark")) 
        val = NULL
    else val = .coerce(val = val, as = argType(argName, expectedArgs))
    return(val)
  }

[Package argparse version 1.0 Index]