parseArguments {argparse}	R Documentation

Invoke argument parser for a given textual description of expected arguments

Description

Top-Level function. Expects a table like this as string:

a | numeric | <> | Help text for a b | character | abc | Help text for b c | logical | FALSE | Help text for c d | numeric | 1 | Help text for d

Usage

parseArguments(expectedArgumentText)

Arguments

expectedArgumentText
String forming a table with 4 columns: name, type, default value or <> if mandatory, Help text.

Value

Nested list of class "expectedArguments". Contains one item per argument, each item containing 4 sub-items named name, type, val, help.

Author(s)

Moritz Lell

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (expectedArgumentText) 
{
    return(parseCommandLine(parseExpectedArgs(expectedArgumentText)))
  }

[Package argparse version 1.0 Index]