parseCommandLine {argparse}	R Documentation

Parse command line using type and default value information

Description

Reads in the command line arguments using argparse. Converts values according to their expected type and fills in default values of missing optional arguments.

Usage

parseCommandLine(expectedArgs)

Arguments

expectedArgs
Description of expected arguments, as parsed by parseExpectedArgs.

Value

List: keys named like expected arguments, values are their values.

Author(s)

Moritz Lell

[Package argparse version 1.0 Index]