parseExpectedArgs {argparse}	R Documentation

Parse textual information about expected variables.

Description

Reads in a string forming a table. Returns a nested list of class "expectedArguments". Contains one item per argument, each item containing 4 sub-items named name, type, val, help.

Usage

parseExpectedArgs(s, sep = "|", expectedMark = "<>", helpText = "")

Arguments

s
String forming a table: Example: a | numeric | <> | Help text for a b | character | abc | Help text for b c | logical | FALSE | Help text for c d | numeric | 1 | Help text for d

sep
Table column separator character

expectedMark
Text in value column if the argument is mandatory. Default: <>.

helpText
A text describing the program. This is shown in addition to the argument specific help texts when the script is invoked with the –help or -h option.

Value

see Description.

Author(s)

Moritz Lell

Examples

expectedArgs="
nreads        | numeric | <> | Number of reads to create
readlength    | numeric | <> | Average length of an individual read
readlength.sd | numeric | <> | Standard deviation of read lengths
start         | numeric | <> | start index
end           | numeric | <> | end index
seed          | numeric | NA | random number generator seed
"

eargs= parseExpectedArgs(expectedArgs)

[Package argparse version 1.0 Index]